# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: StoryShelf API
  version: 1.0.0
paths:
...
  /api/books/:
    get:
      operationId: books_list
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: per_page
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - books
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBookListList'
          description: ''
    post:
      operationId: books_create
      tags:
      - books
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookWrite'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/BookWrite'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/BookWrite'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookWrite'
          description: ''
...
components:
  schemas:
...
    BookList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 500
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
        cover_url:
          type: string
          format: uri
          maxLength: 500
        authors:
          type: array
          items:
            type: string
        genres:
          type: array
          items:
            type: string
        avg_rating:
          type: number
          format: double
        year:
          type: integer
          maximum: 2147483647
          minimum: 1
          nullable: true
      required:
      - authors
      - genres
      - id
      - title
    BookWrite:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        year:
          type: integer
          maximum: 2147483647
          minimum: 1
          nullable: true
        isbn:
          type: string
          nullable: true
          maxLength: 20
        description:
          type: string
        page_count:
          type: integer
          maximum: 2147483647
          minimum: 1
          nullable: true
        cover_url:
          type: string
          format: uri
          maxLength: 500
        serie:
          allOf:
          - $ref: '#/components/schemas/_SerieWrite'
          writeOnly: true
        authors:
          type: array
          items:
            type: string
          writeOnly: true
        genres:
          type: array
          items:
            type: string
          writeOnly: true
        tags:
          type: array
          items:
            type: string
          writeOnly: true
        position_in_series:
          type: integer
          maximum: 2147483647
          minimum: -2147483648
          nullable: true
      required:
      - title
...
    PaginatedBookListList:
      type: object
      required:
      - data
      - page
      - per_page
      - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BookList'
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
...
    _SerieWrite:
      type: object
      properties:
        name:
          type: string
      required:
      - name
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT